requirements.txt
streamlit
sentence-transformers
scikit-learn
numpy
pillow
imagehash
